requirements.txt
Flask
Flask-Compress
rembg
opencv-python-headless
Pillow
numpy